+++ lib/amazing_print/formatters/hash_formatter.rb
@@ -84,8 +84,18 @@ module AmazingPrint
         result
       end
 
+      def key_size(key)
+        return key.inspect.size if symbol?(key)
+
+        if options[:html]
+          single_line { inspector.awesome(key) }.size
+        else
+          plain_single_line { inspector.awesome(key) }.size
+        end
+      end
+
       def max_key_width(keys)
-        keys.map { |key, _value| colorless_size(key.to_s) }.max || 0
+        keys.map { |key, _value| key_size(key.to_s) }.max || 0
       end
 
       def printable_keys
@@ -94,11 +104,8 @@ module AmazingPrint
         keys.sort! { |a, b| a.to_s <=> b.to_s } if options[:sort_keys]
 
         keys.map! do |key|
-          plain_single_line do
-            [
-              json_format? ? key : inspector.awesome(key),
-              hash[key]
-            ]
+          single_line do
+            [key, hash[key]]
           end
         end
       end
@@ -108,7 +115,7 @@ module AmazingPrint
       end
 
       def symbol?(key)
-        key[0] == ':'
+        key.is_a?(Symbol)
       end
 
       def json_format?
@@ -123,16 +130,26 @@ module AmazingPrint
       end
 
       def ruby19_syntax(key, value, width)
-        return pre_ruby19_syntax(key, value, width) unless symbol?(key)
+        # Move the colon to the right side of the symbol
+        awesome_key = inspector.awesome(key).sub(/#{key.inspect}/, "#{key}:")
 
-        "#{align(key[1..-1], width - 1)}#{colorize(': ', :hash)}#{inspector.awesome(value)}"
+        "#{align(awesome_key, width)} #{inspector.awesome(value)}"
       end
 
       def pre_ruby19_syntax(key, value, width)
-        "#{align(key, width)}#{colorize(' => ', :hash)}#{inspector.awesome(value)}"
+        awesome_key = single_line { inspector.awesome(key) }
+        "#{align(awesome_key, width)}#{colorize(' => ', :hash)}#{inspector.awesome(value)}"
       end
 
-      def plain_single_line
+      def plain_single_line(&)
+        plain = options[:plain]
+        options[:plain] = true
+        single_line(&)
+      ensure
+        options[:plain] = plain
+      end
+
+      def single_line
         multiline = options[:multiline]
         options[:multiline] = false
         yield