+++ spec/ext/active_model_spec.rb
@@ -3,9 +3,32 @@
 require 'spec_helper'
 require 'active_record_helper'
 
-RSpec.describe 'ActiveModel::Errors formatting', skip: -> { !ExtVerifier.has_rails? }.call do
+RSpec.describe 'ActiveModel formatting', skip: -> { !ExtVerifier.has_rails? }.call do
   before do
     @ap = AmazingPrint::Inspector.new(plain: true)
+  end
+
+  describe 'ActiveModel class' do
+    it 'prints the class' do
+      expect(@ap.awesome(ActiveModelModel)).to eq <<~PRINT.strip
+        class ActiveModelModel < Object {
+            :name => :string,
+            :rank => :integer
+        }
+      PRINT
+    end
+  end
+
+  describe 'ActiveModel instance' do
+    it 'prints the instance' do
+      model = ActiveModelModel.new(name: 'John', rank: 1)
+      expect(@ap.awesome(model)).to be_similar_to <<~PRINT.strip
+        #<ActiveModelModel:placeholder_id> {
+            :name => "John",
+            :rank => 1
+        }
+      PRINT
+    end
   end
 
   it 'formats active_model_errors properly' do